Easy Creature Blender Addon
Automates adding Mirror, Skin, and Subdivision modifiers for organic modeling. Best used with simple starting geometries, such as a single vertex, to build characters.
Warning: Not recommended for use on complex models with a lot of geometries. The load may be too much for your system and may crash Blender.